pyproject.toml
[project]
name = "django-filter-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.13.1",
]
